+++ db.py
@@ -35,8 +35,6 @@ def init_db():
     db = get_db()
 
     with current_app.open_resource('schema.sql') as f:
-        #
-    #
         db.executescript(f.read().decode('utf8'))
 
 @click.command('init-db')
@@ -47,6 +45,5 @@ def init_db_command():
     click.echo('Initialized the database.')
 
 def init_app(app):
-    
     app.teardown_appcontext(close_db)
     app.cli.add_command(init_db_command)
