+++ requirements.txt
@@ -1,0 +1,6 @@
+click==8.1.3
+Flask==2.1.2
+itsdangerous==2.1.2
+Jinja2==3.1.2
+MarkupSafe==2.1.1
+Werkzeug==2.1.2
